Refactor: extract server, db, resources, ... replace them with import from the new module of the platform and keep only classes that handle OPA server and its resource
