+++ opa-client/src/main/java/org/opendatamesh/platform/up/policy/opa/client/resources/v1/ValidateRequest.java
@@ -1,0 +1,10 @@
+package org.opendatamesh.platform.up.policy.opa.client.resources.v1;
+
+import lombok.Data;
+
+@Data
+public class ValidateRequest {
+
+  private Object input;
+  
+}
