--- opa-client/src/main/java/org/opendatamesh/policyserviceopa/client/OpaRestTemplate.java
@@ -1,25 +1,0 @@
-package org.opendatamesh.policyserviceopa.client;
-
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.web.client.RestTemplate;
-
-import java.time.Duration;
-
-public class OpaRestTemplate {
-
-    Long timeout;
-
-    RestTemplateBuilder restTemplateBuilder = new RestTemplateBuilder();
-
-    public OpaRestTemplate(Long timeout) {
-        this.timeout = timeout;
-    }
-
-    public RestTemplate buildRestTemplate() {
-        return restTemplateBuilder
-                .setConnectTimeout(Duration.ofSeconds(timeout))
-                .setReadTimeout(Duration.ofSeconds(timeout))
-                .build();
-    }
-
-}
