--- opa-client/src/main/java/org/opendatamesh/policyserviceopa/client/resources/v1/ValidateRequest.java
@@ -1,10 +1,0 @@
-package org.opendatamesh.policyserviceopa.client.resources.v1;
-
-import lombok.Data;
-
-@Data
-public class ValidateRequest {
-
-  private Object input;
-  
-}
