--- opa-policy-server/src/main/java/org/opendatamesh/policyserviceopa/server/PolicyserviceOpaApplication.java
@@ -1,13 +1,0 @@
-package org.opendatamesh.policyserviceopa.server;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PolicyserviceOpaApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PolicyserviceOpaApplication.class, args);
-	}
-
-}
