--- opa-policy-server/src/main/java/org/opendatamesh/policyserviceopa/server/entities/PolicyEntity.java
@@ -1,40 +1,0 @@
-package org.opendatamesh.policyserviceopa.server.entities;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import java.sql.Timestamp;
-
-@Entity
-@Table(name="POLICY")
-@Data
-public class PolicyEntity {
-    @Id
-    @Column(name = "ID", updatable = false, nullable = false)
-    private String id;
-
-    @Column(name = "DISPLAY_NAME")
-    private String displayName;
-
-    @Column(name = "DESCRIPTION")
-    private String description;
-
-    @Column(name = "RAW_POLICY")
-    private String rawPolicy;
-
-    @Column(name = "CREATED_AT", updatable = false, nullable = false)
-    @CreationTimestamp
-    private Timestamp createdAt;
-
-    @Column(name = "UPDATED_AT")
-    @UpdateTimestamp
-    private Timestamp updatedAt;
-
-}
