--- opa-policy-server/src/main/java/org/opendatamesh/policyserviceopa/server/entities/SuiteEntity.java
@@ -1,36 +1,0 @@
-package org.opendatamesh.policyserviceopa.server.entities;
-
-import lombok.Data;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import javax.persistence.*;
-import java.sql.Timestamp;
-import java.util.List;
-
-@Entity
-@Table(name="SUITE")
-@Data
-public class SuiteEntity {
-    @Id
-    @Column(name = "ID", updatable = false, nullable = false)
-    private String id;
-
-    @Column(name = "DISPLAY_NAME")
-    private String displayName;
-
-    @Column(name = "DESCRIPTION")
-    private String description;
-
-    @Column(name = "CREATED_AT", updatable = false, nullable = false)
-    @CreationTimestamp
-    private Timestamp createdAt;
-
-    @Column(name = "UPDATED_AT")
-    @UpdateTimestamp
-    private Timestamp updatedAt;
-
-    @ElementCollection
-    private List<String> policies;
-
-}
