--- opa-policy-server/src/main/resources/db/migration/mysql/V1.0__init.sql
@@ -1,25 +1,0 @@
--- MySql 8
-
-CREATE TABLE POLICY (
-    ID VARCHAR(255) NOT NULL,
-    DISPLAY_NAME VARCHAR(255),
-    DESCRIPTION VARCHAR(1000),
-    RAW_POLICY BLOB,
-    CREATED_AT TIMESTAMP,
-    UPDATED_AT TIMESTAMP,
-    PRIMARY KEY (ID)
-);
-
-CREATE TABLE SUITE (
-    ID VARCHAR(255),
-    DISPLAY_NAME VARCHAR(255),
-    DESCRIPTION VARCHAR(1000),
-    CREATED_AT TIMESTAMP,
-    UPDATED_AT TIMESTAMP,
-    PRIMARY KEY (ID)
-);
-
-CREATE TABLE SUITEENTITY_POLICIES (
-    SUITEENTITY_ID VARCHAR(255) NOT NULL,
-    POLICIES VARCHAR(255)
-);
