--- opa-policy-server/src/main/resources/db/migration/postgres/V1.0__init.sql
@@ -1,25 +1,0 @@
--- Postgres 11
-
-CREATE SCHEMA IF NOT EXISTS "PUBLIC";
-
-CREATE TABLE "PUBLIC"."POLICY"(
-    "ID" VARCHAR(255) PRIMARY KEY,
-    "DISPLAY_NAME" VARCHAR(255),
-    "DESCRIPTION" VARCHAR(1000),
-    "RAW_POLICY" TEXT,
-    "CREATED_AT" TIMESTAMP,
-    "UPDATED_AT" TIMESTAMP
-);
-
-CREATE TABLE "PUBLIC"."SUITE"(
-    "ID" VARCHAR(255) PRIMARY KEY,
-    "DISPLAY_NAME" VARCHAR(255),
-    "DESCRIPTION" VARCHAR(1000),
-    "CREATED_AT" TIMESTAMP,
-    "UPDATED_AT" TIMESTAMP
-);
-
-CREATE TABLE "PUBLIC"."SuiteEntity_policies"(
-    "SuiteEntity_ID" VARCHAR(255) NOT NULL,
-    "policies" VARCHAR(255)
-);
